+++ CHANGES.rst
@@ -4,6 +4,12 @@ Changelog
 Here you can see the full list of changes between each SQLAlchemy-JSON-API release.
 
 
+0.4.3 (2017-07-28)
+^^^^^^^^^^^^^^^^^^
+
+- Fixed column adaptation for hybrid / column properties returning Column objects
+
+
 0.4.2 (2017-03-17)
 ^^^^^^^^^^^^^^^^^^